+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - feat!: All order book events now carry the relevant data for the event. And events that are not related to offers do not carry offer data.
 - feat: The market order simulation used to estimate volumes and gas has been updated to match Mangrove v2's market order logic.
 - feat!: 'Market.estimateVolume' now also estimates fees and returns it in a new 'estimatedFee' field. The existing 'estimatedVolume' field is exclusive of fees and thus represents the true amount the taker can expect to receive/pay.
+- feat: Add 'Market.retractRestingOrder' method for retracting resting orders posted by 'MangroveOrder'.
 
 # 2.0.0-12